Updated and added new docs.
Changes to refactored grammar to better match it's implementation in code. Same can be said for the first and follows sets as well. The state table is an addition for the updated scanner that now uses the state design pattern. When encountering a character in a Graphene file the table shows what state to proceed to in order to make the token or if the token can not be made a dead state has been reached and we can terminate the scan. The parsing table is similar to the state table in that it gives the next rules to follow, however, these rules follow the productions of non-terminals in the refactored grammar until terminal values are reached, or are not reached, in which case the parse can be terminated.